Add an intermediate mount for a ploop
We need to know a mount point to create a snapshot. We can parse
/proc/self/mountinfo and try to find this mount, but it is easier
to mount it into a known place.

In addition, this can help us to support a case, when we need to mount
one ploop into a few pods (e.g. ReadOnlyMany). Ploop.Mount() can't be
called a few times for one ploop, but if we will know a mount point, we
will be able to bind-mount it to somewere else.