Add a hook to call a local ploop if available
Right now we have a situation, that for k8s snapshots feature we need a more
recent ploop, rather then ploop in standalone virtuozzo storage product. This
situation is a typical case, when there are several product with different
requirements use the same component, but conflicts by version.

Thus, it is better to have a hook in ploop-flexvol code, which allows to use
custom ploop just for it, if necessary.

Signed-off-by: Dmitry Mishin <dim@virtuozzo.com>